flask-server/requirement.txt
langchain==0.1.8
langchain_community==0.0.21
langchain_openai==0.0.6
openai==1.12.0
chromadb==0.3.29
flask==3.0.0
flask_cors==4.0.0
python-dotenv==1.0.0
pypdf==3.17.0
tiktoken==0.6.0
jinja2==3.1.2


